Added tests for filter render
