+++ src/component/__test__/ComicDetail.test.jsx
@@ -1,0 +1,3 @@
+test('', ()=>{
+  expect('0'==='0');
+})
