+++ src/testUtils/dataMock.js
@@ -1,0 +1,45 @@
+const data = [ 
+    {
+      "id": 58587,
+      "title": "The Amazing Spider-Man (2015) #21 (Rivera Variant)",
+      "issueNumber": 21,
+      "description": null,
+      "format": "Comic",
+      "urls": [
+        {
+          "type": "detail",
+          "url": "http://marvel.com/comics/issue/58587/the_amazing_spider-man_2015_21_rivera_variant/rivera_variant?utm_campaign=apiRef&utm_source=90c704a85bddb1baa24fcd3058cc5d3c"
+          }
+        ],
+      "thumbnail": {
+        "path": "http://i.annihil.us/u/prod/marvel/i/mg/b/40/image_not_available",
+        "extension": "jpg"
+      },
+      "images": []
+    },
+    {
+      "id": 40921,
+      "title": "Ghost Rider: Official Index to the Marvel Universe GN-TPB (Graphic Novel)",
+      "issueNumber": 1,
+      "description": "The complete history of the Spirit of Vengeance from\r\nhis earliest appearances to the present day! This\r\nfact-packed volume chronicles every character, team,\r\nplace and piece of equipment &mdash; and provides vital\r\ninformation about all things Ghost Rider! Collecting\r\nthe Ghost Rider material from WOLVERINE, PUNISHER\r\n& GHOST RIDER: OFFICIAL INDEX TO THE MARVEL\r\nUNIVERSE #1-6.\r\n152 PGS./Rated T+ &#133;$19.99",
+      "format": "Graphic Novel",
+      "urls": [
+        {
+          "type": "detail",
+          "url": "http://marvel.com/comics/collection/40921/ghost_rider_official_index_to_the_marvel_universe_gn-tpb_graphic_novel?utm_campaign=apiRef&utm_source=90c704a85bddb1baa24fcd3058cc5d3c"
+          }
+        ],
+      "thumbnail": {
+        "path": "http://i.annihil.us/u/prod/marvel/i/mg/b/a0/4ec2b1272cee0",
+        "extension": "jpg"
+      },
+      "images": [
+        {
+          "path": "http://i.annihil.us/u/prod/marvel/i/mg/b/a0/4ec2b1272cee0",
+          "extension": "jpg"
+          }
+        ],
+      }
+ ];
+
+export default data;
